requirements.txt
pytest
selenium
pylint
python-dotenv